requirements.txt
blinker==1.6.2
click==8.1.7
dnspython==2.4.2
Flask==2.3.2
Flask-Login==0.6.2
Flask-PyMongo==2.3.0
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
neo4j==5.11.0
packaging==23.1
pymongo==4.4.1
python-dotenv==1.0.0
pytz==2023.3
Werkzeug==2.3.7
